refactor: move pull/push to config subcommands, remove unused cache/data dirs, update docs
- Reorganize pull/push as config subcommands following Cobra best practices
- Remove unused .anvil/cache and .anvil/data directories
- Update all documentation and fix imports
