+++ docs/CONTRIBUTING.md
@@ -99,8 +99,8 @@ anvil/
 │   ├── setup/             # Setup command
 │   ├── draw/              # Draw command
 │   ├── config/            # Config command (parent for pull/push)
-│   ├── pull/              # Pull subcommand (config pull)
-│   ├── push/              # Push subcommand (config push)
+│   │   ├── pull/          # Pull subcommand (config pull)
+│   │   └── push/          # Push subcommand (config push)
 │   └── root.go            # Root command configuration
 ├── pkg/                   # Reusable packages
 │   ├── brew/              # Homebrew integration
